Updated methods section in the .Rmd file from comments in the google doc.
There's a major change to how we describe the number of RILs used in each experiment.
I realised that this was redundant - we actually use almost all the RILs, but the text still said we were using as few as 338 RILs for certain traits.
Updated to focus on differences with the previous paper from 2013.